Update iPhoneAssignmentsRxTrack
Select Class: The Select class is specifically designed for interacting with <select> HTML elements. It allows you to select options by visible text, value, or index.
select_by_visible_text: This method is used to select an option by the text that is visible to the user. In your case, it would be "ATC PREMEDS Tracking".